chore: move bounding box debug rendering code to dedicated BoundingBoxRenderer
