+++ engine/src/main/java/org/terasology/engine/physics/engine/PhysicsSystem.java
@@ -3,15 +3,11 @@
 package org.terasology.engine.physics.engine;
 
 import com.google.common.collect.Lists;
-import org.joml.Matrix4f;
 import org.joml.Quaternionf;
 import org.joml.Vector3f;
-import org.lwjgl.opengl.GL33;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.engine.config.Config;
 import org.terasology.engine.core.Time;
-import org.terasology.engine.entitySystem.entity.EntityManager;
 import org.terasology.engine.entitySystem.entity.EntityRef;
 import org.terasology.engine.entitySystem.entity.lifecycleEvents.BeforeDeactivateComponent;
 import org.terasology.engine.entitySystem.entity.lifecycleEvents.OnActivatedComponent;
@@ -21,7 +17,6 @@ import org.terasology.engine.entitySystem.event.ReceiveEvent;
 import org.terasology.engine.entitySystem.systems.BaseComponentSystem;
 import org.terasology.engine.entitySystem.systems.RegisterMode;
 import org.terasology.engine.entitySystem.systems.RegisterSystem;
-import org.terasology.engine.entitySystem.systems.RenderSystem;
 import org.terasology.engine.entitySystem.systems.UpdateSubscriberSystem;
 import org.terasology.engine.logic.location.LocationComponent;
 import org.terasology.engine.logic.location.LocationResynchEvent;
@@ -33,10 +28,6 @@ import org.terasology.engine.physics.HitResult;
 import org.terasology.engine.physics.StandardCollisionGroup;
 import org.terasology.engine.physics.components.RigidBodyComponent;
 import org.terasology.engine.physics.components.TriggerComponent;
-import org.terasology.engine.physics.components.shapes.BoxShapeComponent;
-import org.terasology.engine.physics.components.shapes.CapsuleShapeComponent;
-import org.terasology.engine.physics.components.shapes.CylinderShapeComponent;
-import org.terasology.engine.physics.components.shapes.SphereShapeComponent;
 import org.terasology.engine.physics.events.BlockImpactEvent;
 import org.terasology.engine.physics.events.ChangeVelocityEvent;
 import org.terasology.engine.physics.events.CollideEvent;
@@ -46,20 +37,10 @@ import org.terasology.engine.physics.events.ImpactEvent;
 import org.terasology.engine.physics.events.ImpulseEvent;
 import org.terasology.engine.physics.events.PhysicsResynchEvent;
 import org.terasology.engine.registry.In;
-import org.terasology.engine.rendering.assets.material.Material;
-import org.terasology.engine.rendering.assets.mesh.Mesh;
-import org.terasology.engine.rendering.assets.mesh.StandardMeshData;
-import org.terasology.engine.rendering.assets.mesh.resource.AllocationType;
-import org.terasology.engine.rendering.assets.mesh.resource.DrawingMode;
-import org.terasology.engine.rendering.world.WorldRenderer;
-import org.terasology.engine.utilities.Assets;
 import org.terasology.engine.world.OnChangedBlock;
 import org.terasology.engine.world.WorldProvider;
 import org.terasology.engine.world.block.Block;
 import org.terasology.engine.world.block.BlockComponent;
-import org.terasology.gestalt.assets.management.AssetManager;
-import org.terasology.joml.geom.AABBf;
-import org.terasology.nui.Color;
 
 import java.util.Iterator;
 import java.util.List;
@@ -72,7 +53,7 @@ import java.util.List;
  *
  */
 @RegisterSystem
-public class PhysicsSystem extends BaseComponentSystem implements UpdateSubscriberSystem, RenderSystem {
+public class PhysicsSystem extends BaseComponentSystem implements UpdateSubscriberSystem {
 
     private static final Logger logger = LoggerFactory.getLogger(PhysicsSystem.class);
     private static final long TIME_BETWEEN_NETSYNCS = 500;
@@ -87,26 +68,11 @@ public class PhysicsSystem extends BaseComponentSystem implements UpdateSubscrib
     private PhysicsEngine physics;
     @In
     private WorldProvider worldProvider;
-    @In
-    protected AssetManager assetManager;
-    @In
-    protected WorldRenderer worldRenderer;
-    @In
-    protected Config config;
-    @In
-    protected EntityManager entityManager;
-
-    private StandardMeshData meshData = new StandardMeshData(DrawingMode.LINES, AllocationType.STREAM);
-    private Mesh mesh;
-    private Material material;
 
     private long lastNetsync;
 
     @Override
     public void initialise() {
-        material = assetManager.getAsset("engine:white", Material.class).get();
-        mesh = Assets.generateAsset(meshData, Mesh.class);
-
         lastNetsync = 0;
     }
 
@@ -192,123 +158,6 @@ public class PhysicsSystem extends BaseComponentSystem implements UpdateSubscrib
             rigidBody.setLocation(vNewPosition);
             rigidBody.setLinearVelocity(vNewVelocity);
             rigidBody.setAngularVelocity(vNewVelocity);
-        }
-    }
-
-    private int addRenderBound(StandardMeshData meshData, AABBf bounds, int index) {
-        Vector3f pos = new Vector3f();
-        meshData.position.put(pos.set(bounds.minX, bounds.minY, bounds.minZ));
-        meshData.position.put(pos.set(bounds.maxX, bounds.minY, bounds.minZ));
-        meshData.position.put(pos.set(bounds.maxX, bounds.minY, bounds.maxZ));
-        meshData.position.put(pos.set(bounds.minX, bounds.minY, bounds.maxZ));
-
-        meshData.position.put(pos.set(bounds.minX, bounds.maxY, bounds.minZ));
-        meshData.position.put(pos.set(bounds.maxX, bounds.maxY, bounds.minZ));
-        meshData.position.put(pos.set(bounds.maxX, bounds.maxY, bounds.maxZ));
-        meshData.position.put(pos.set(bounds.minX, bounds.maxY, bounds.maxZ));
-
-        meshData.color0.put(Color.black);
-        meshData.color0.put(Color.black);
-        meshData.color0.put(Color.black);
-        meshData.color0.put(Color.black);
-
-        meshData.color0.put(Color.black);
-        meshData.color0.put(Color.black);
-        meshData.color0.put(Color.black);
-        meshData.color0.put(Color.black);
-
-        meshData.indices.putAll(new int[]{
-                // top loop
-                index, index + 1,
-                index + 1, index + 2,
-                index + 2, index + 3,
-                index + 3, index,
-
-                // connecting edges between top and bottom
-                index, index + 4,
-                index + 1, index + 5,
-                index + 2, index + 6,
-                index + 3, index + 7,
-
-                // bottom loop
-                index + 4, index + 5,
-                index + 5, index + 6,
-                index + 6, index + 7,
-                index + 7, index + 4,
-
-        });
-        return index + 8;
-    }
-
-    @Override
-    public void renderOverlay() {
-        if (config.getRendering().getDebug().isRenderingEntityColliders()) {
-            GL33.glDepthFunc(GL33.GL_ALWAYS);
-            Vector3f cameraPosition = worldRenderer.getActiveCamera().getPosition();
-
-            Vector3f worldPos = new Vector3f();
-            Vector3f worldPositionCameraSpace = new Vector3f();
-            worldPos.sub(cameraPosition, worldPositionCameraSpace);
-            Matrix4f matrixCameraSpace = new Matrix4f().translationRotateScale(worldPositionCameraSpace, new Quaternionf(), 1.0f);
-            Matrix4f modelViewMatrix = new Matrix4f(worldRenderer.getActiveCamera().getViewMatrix()).mul(matrixCameraSpace);
-            material.setMatrix4("projectionMatrix", worldRenderer.getActiveCamera().getProjectionMatrix());
-            material.setMatrix4("modelViewMatrix", modelViewMatrix, true);
-
-            int index = 0;
-            meshData.reallocate(0, 0);
-            meshData.indices.rewind();
-            meshData.position.rewind();
-            meshData.color0.rewind();
-
-            Vector3f worldPosition = new Vector3f();
-            Quaternionf worldRot = new Quaternionf();
-            Matrix4f transform = new Matrix4f();
-            AABBf bounds = new AABBf(0, 0, 0, 0, 0, 0);
-
-            for (EntityRef entity : entityManager.getEntitiesWith(LocationComponent.class)) {
-                LocationComponent location = entity.getComponent(LocationComponent.class);
-                location.getWorldPosition(worldPosition);
-                location.getWorldRotation(worldRot);
-
-                BoxShapeComponent boxShapeComponent = entity.getComponent(BoxShapeComponent.class);
-                if (boxShapeComponent != null) {
-                    bounds.set(0, 0, 0, 0, 0, 0);
-                    bounds.expand(new Vector3f(boxShapeComponent.extents).div(2.0f));
-                    transform.translationRotateScale(worldPosition, worldRot, location.getWorldScale());
-                    bounds.transform(transform);
-                    index = addRenderBound(meshData, bounds, index);
-                }
-                CapsuleShapeComponent capsuleComponent = entity.getComponent(CapsuleShapeComponent.class);
-                if (capsuleComponent != null) {
-                    bounds.set(0, 0, 0, 0, 0, 0);
-                    bounds.expand(new Vector3f(capsuleComponent.radius, capsuleComponent.height / 2.0f, capsuleComponent.radius).div(2.0f));
-                    transform.translationRotateScale(worldPosition, worldRot, location.getWorldScale());
-                    bounds.transform(transform);
-                    index = addRenderBound(meshData, bounds, index);
-                }
-                CylinderShapeComponent cylinderShapeComponent = entity.getComponent(CylinderShapeComponent.class);
-                if (cylinderShapeComponent != null) {
-                    bounds.set(0, 0, 0, 0, 0, 0);
-                    bounds.expand(new Vector3f(cylinderShapeComponent.radius, cylinderShapeComponent.height / 2.0f,
-                            cylinderShapeComponent.radius).div(2.0f));
-                    transform.translationRotateScale(worldPosition, worldRot, location.getWorldScale());
-                    bounds.transform(transform);
-                    index = addRenderBound(meshData, bounds, index);
-                }
-                SphereShapeComponent sphereShapeComponent = entity.getComponent(SphereShapeComponent.class);
-                if (sphereShapeComponent != null) {
-                    bounds.set(0, 0, 0, 0, 0, 0);
-                    bounds.expand(new Vector3f(sphereShapeComponent.radius).div(2.0f));
-                    transform.translationRotateScale(worldPosition, worldRot, location.getWorldScale());
-                    bounds.transform(transform);
-                    index = addRenderBound(meshData, bounds, index);
-                }
-            }
-
-            material.enable();
-            mesh.reload(meshData);
-            mesh.render();
-            GL33.glDepthFunc(GL33.GL_LEQUAL);
         }
     }
